+++ src/main/java/net/soulsweaponry/config/ClientConfig.java
@@ -1,0 +1,8 @@
+package net.soulsweaponry.config;
+
+public class ClientConfig extends MidnightConfig {
+
+    @Client @Entry public static boolean disable_target_posture_hud = false;
+    @Client @Entry public static boolean disable_player_posture_hud = false;
+    @Client @Entry public static boolean disable_player_bleed_hud = false;
+}