+++ src/main/java/net/soulsweaponry/events/PlayerTickHandler.java
@@ -5,6 +5,7 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.soulsweaponry.entitydata.IEntityDataSaver;
 import net.soulsweaponry.entitydata.ParryData;
+import net.soulsweaponry.entitydata.TargetPostureData;
 
 public class PlayerTickHandler implements ServerTickEvents.StartTick {
 
@@ -15,6 +16,11 @@ public class PlayerTickHandler implements ServerTickEvents.StartTick {
             if (frames >= 1) {
                 ParryData.addParryFrames((IEntityDataSaver) player, 1);
                 player.stopUsingItem();
+            }
+            if (player.getAttacking() != null) {
+                TargetPostureData.updateTargetPosture((IEntityDataSaver) player, player.getAttacking());
+            } else {
+                TargetPostureData.resetValues((IEntityDataSaver) player);
             }
         }
     }