+++ src/main/java/net/soulsweaponry/networking/PacketRegistry.java
@@ -34,5 +34,6 @@ public class PacketRegistry {
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.STOP_BOSS_MUSIC, StopBossMusicS2C::receive);
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.CHAIN_LIGHTNING, ChainLightningS2C::receive);
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.BLEED_SYNC, BleedSyncS2C::receive);
+        ClientPlayNetworking.registerGlobalReceiver(PacketIds.TARGET_POSTURE_SYNC, TargetPostureSyncS2C::receive);
     }
 }